backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
razorpay>=1.0
requests>=2.31
